python/hosted_agents/pyproject.toml
[project]
name = "hosted-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-agentserver-agentframework>=1.0.0b13",
    "azure-ai-agentserver-core>=1.0.0b13",
    "azure-ai-projects>=2.0.0",
    "azure-identity>=1.26.0b2",
    "jupyter>=1.1.1",
    "python-dotenv>=1.2.2",
]
